pyproject.toml
[project]
name = "markedsafeloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]
